gnu: Deprecate gnome meta-package.
GNOME is not a "one size fits all" environment, but with the introduction of
World, Circle, and to some extent Developer Extras and recent explorations of
mobile form factors much more customizable towards the needs of users.  As
such, it is silly to insist on a single meta-package to make everyone happy.

* gnu/packages/gnome.scm (gnome): Deprecate.