gnu: python-pyportmidi: Update to 0.0.7-0.d9e5ee0.
When the PortMidi project moved to GitHub in 2021, changing version numbering
scheme, the Python bindings were moved to a separate repository and marked as
needing a new maintainer.  The chosen revision, dated June 2023, is the first
to allow the package to build successfully in the temporary lack of an
official new release.

* gnu/packages/music.scm (python-pyportmidi): Source moved, fix build.
[version]: Change to new scheme, jumping from 217 to 0.0.7.
[source]: Change origin from Sourceforge to GitHub.

Change-Id: Ibe02edca39731f3f3c9001818665d2da97df9465
Signed-off-by: Christopher Baines <mail@cbaines.net>