gnu: notmuch: Add sfsexp input.
sfsexp is a C/C++ library for the manipulation (read, parse, modify, and
create) of symbolic expressions.  In an environment where sfsexp is available,
Notmuch will compile with support for a S-expression-based query format.  This
commit is to add sfsexp as a Notmuch dependency and therefore to add S-exp
support.

* gnu/packages/mail.scm (notmuch)[native-inputs]: Add git-minimal/pinned.
[inputs]: Add sfsexp.

Change-Id: Ic3b4a30b2672ad690d66b283e9bdc8e2197f97a7
Signed-off-by: Christopher Baines <mail@cbaines.net>