gnu: musescore: Propagate alsa-plugins pulseaudio to runtime.
In order for libasound_module_conf_pulse.so to be present at runtime of
MuseScore the alsa-plugins output "pulseaudio" must be present.

Playback of scores was not working, at least on a foreign distro like
Debian. When running on Debian, MuseScore complained that
libasound_module_conf_pulse.so was missing at startup. This patch was
tested on Debian 12 and fixed the issue.

* gnu/packages/music.scm (musescore): Propagate alsa-plugins pulseaudio to runtime

Signed-off-by: Nicolas Goaziou <mail@nicolasgoaziou.fr>
Change-Id: I4e2db8e45fdf33987817cef41bcd336526b16d90