gnu: Add emacs-helm-org-ql.
* gnu/packages/emacs-xyz.scm (emacs-helm-org-ql): New variable.
(emacs-org-ql)[phases]: Add phase 'remove-helm-org-ql.
[propagated-inputs]: Remove emacs-helm, emacs-helm-org.

The author didn't mean to make the package depend on helm :
"Note that Helm is not declared as a package dependency, so this does
not cause Helm to be installed.", but the way we packaged it in Guix
did. Split the extension into a separate dependency instead, this
avoids 3 Mo to be propagated in a profile where it's not meant.

Change-Id: I3537c0b08cfffac3d66b761173761a3806497358
Signed-off-by: Andrew Tropin <andrew@trop.in>