committer: Avoid // in xpath.
When using // we assume that the package definition's S-expression is a
nesting of proper lists.  This is not guaranteed to be true, because a package
definition may contain alists.  By using an explicit path to the field of
interest we avoid recursion through all child nodes, thus avoiding child nodes
that are more likely to contain improper lists.

* etc/committer.scm.in (change-commit-message): Replace // with a path for the
PACKAGE symbol.

Change-Id: Idefaec9f5c3538484432e15a5865a68fc16e7369