gnu: guix-daemon: Remove broken 'change-default-guix' phase.
It has no effect, probably due to the '+' in the pattern not being escaped.
We also don't want to refer to /var/guix/profiles/per-user/root/bin/guix
anymore, which may not exist (for Guix System users that have never ran 'guix
pull' has root).  The 'guix' command used is that specified via the build
system, which is $(bindir)/guix.

* gnu/packages/package-management.scm (guix-daemon) [phases]: Delete
change-default-guix phase.

Change-Id: If4703334f34aac3486ff655af7d8d4c2ccf0a0eb