gnu: newlib arm-none-eabi-7-2018-q2-update: Add proper newlib-nano variant.
The nano variant for was not nano. It was just a copy of the original
one with different name. This meant there were no _nano.a files present
at all, and the flags were the same, not producing a smaller library.

* gnu/packages/embedded.scm (make-base-newlib-arm-none-eabi-7-2018-q2-update):
Rename variable from make-newlib-arm-none-eabi-7-2018-q2-update.
* gnu/packages/embedded.scm (make-newlib-nano-arm-none-eabi-7-2018-q2-update):
Add variable.
* gnu/packages/embedded.scm (make-newlib-arm-none-eabi-7-2018-q2-update):
Inherit from make-base-newlib-arm-none-eabi-7-2018-q2-update.

Change-Id: I01517ed860dfb2b6a9c7e912fd724832f5a0500a
Signed-off-by: Ludovic Courtès <ludo@gnu.org>