gnu: librewolf-source: Turn into a procedure.
This patch changes the `librewolf-source' variable into the
`make-librewolf-source' prodecure.

This procedure accepts a LibreWolf version, source hash, and Firefox source
hash.  The Firefox source version is derived from the provided LibreWolf
version.

This eases package updates, since the hashes are inside the `librewolf'
package, rather than `librewolf-source'; and the version no longer needs to be
specified in three places.

It also removes a blank line between the file header and `define-module'.

* gnu/packages/librewolf.scm (librewolf-source): Turn into a procedure.

Change-Id: I96ab1304acde246c179e7aa5dad9ff621be3de82
Signed-off-by: Andrew Tropin <andrew@trop.in>