gnu: Update otpclient to 4.0.2
Requires updating `libcotp` to 3.1.0, that introduced a new way for
testing that requires `criterium`, a package we don't have and that has
dependencies that we don't have either.

As the tests in `libcotp` are not an exhaustive testsuite and
`otpclient` is its only dependent package, tests are disabled for the
moment.

* gnu/packages/authentication.scm (libcotp): Update to 3.1.0
[arguments] Disable tests.
* gnu/packages/password-utils.scm (otpclient): Update to 4.0.2

Change-Id: Ie6d37b5efe59397d4979c7490b94f1c871c27d6d