gnu: maturin: Update to 1.8.0.
* gnu/packages/rust-apps.scm (maturin): Update to 1.8.0.
[source]: Remove snippet.
[arguments]: Don't run the tests in release mode.  Adjust the custom
'install-completions phase to install the bash completions in the
correct directory and to install the nushell completions.
[cargo-inputs]: Add rust-pretty-assertions-1, rust-same-file-1,
rust-schemars-0.8, rust-xid-unicode-0.2.  Replace
rust-cargo-metadata-0.18 with 0.19, rust-cbindgen-0.26 with 0.27,
rust-clap-complete-command-0.5 with 0.6, rust-fs-err-2 with 3,
rust-goblin-0.7 with 0.9, rust-minijinja-1 with 2, rust-pep440-rs-0.3
with 0.6, rust-pep508-0.2 with 0.6, rust-pyproject-toml-0.8 with 0.11,
rust-rustls-0.21 with 0.23, rust-target-lexicon-0.12 with 0.13,
rust-thiserror-1 with 2, rust-toml-edit-0.21 with 0.22, rust-zip-0.6
with 2.
[cargo-development-inputs]: Add rust-fs4-0.12, rust-rstest-0.22.
Replace rust-trycmd-0.14 with 0.15, rust-which-5 with 7.
* gnu/packages/patches/maturin-no-cross-compile.patch: Update patch.

Change-Id: Iad445c26556efe4527085876a759d65d4f72621d