gnu: emacs-parinfer-rust-mode: Corrected missing path separator in subsitution
* gnu/packages/emacs-xyz.scm (emacs-parinfer-rust-mode): Corrected missing
path separator in substitution.

The absence of this '/' meant that the parinfer Emacs layer ended up with a
Rust library path that looked like
`/gnu/store/hashvalue-parinfer-rust-emacs-0.4.6/liblibparinfer_rust.so', which
of course doesn't exist.  This patch corrects this behaviour and allows the
Emacs layer to properly find the Rust library.

Change-Id: Idfb5d73e62a7233c133ddc0a53afba5e9de06989
Signed-off-by: Nicolas Goaziou <mail@nicolasgoaziou.fr>