gnu: Add python-numpy-2.
Test suite for NumPy@2.2.2 is heavily depends on included vendored-meson
sub-package and all of the tests fail.  This variant provides a working
NumPy, with passed build and sanity-check phases and tests disabled; it
allows users to include it in other package(s) where version 2+ is hard
required e.g. SageMath or Astropy@7.

* gnu/packages/python-xyz.scm (python-numpy-2): New variable.

Reviewed-by: Ludovic Courtès <ludo@gnu.org>
Change-Id: I07e7aac8c1bf7e686b1c543a11f9ecbf682d6ad5