gnu: libtorrent-rasterbar: Remove special case for test_ssl.
With the release of 2.0.11, the SSL certificates used for tests are valid for
long enough time (~end of year 2297) to no longer require faketime.  With the
removal of faketime, the explicit timeout is also no longer required, since it
does not hang anymore.

This gives us back the validation that SSL works.

* gnu/packages/bittorrent.scm
(libtorrent-rasterbar)[arguments]<#:phases>['check]: Remove test timeout.  Do
not special case test_ssl.

Change-Id: I535c72fec24658a4b2151d2e8794319055c9a278
Signed-off-by: Maxim Cournoyer <maxim.cournoyer@gmail.com>